gestione errori

+++ code/server/src/controllers/userController.ts
@@ -1,5 +1,6 @@
 import { User } from "../components/user";
 import UserDAO from "../dao/userDAO";
+import { UserAlreadyExistsError, UserNotFoundError, UserNotAdminError, UnauthorizedUserError, UserIsAdminError, InvalidBirthdateError, UnauthorizedEditError } from "../errors/userError";
 
 /**
  * Represents a controller for managing users.
@@ -51,7 +52,14 @@ class UserController {
      * @returns A Promise that resolves to the user with the specified username.
      */
     async getUserByUsername(user: User, username: string) /*:Promise<User>*/ {
-        return this.dao.getUserByUsername(username)
+        try {
+            if(user.role !== "Admin" && user.username !== username) {
+                throw new UnauthorizedUserError();
+            }
+            return this.dao.getUserByUsername(username)
+        } catch (error) {
+            throw error;
+        }
     }
      
 
@@ -65,12 +73,24 @@ class UserController {
      */
     async deleteUser(user: User, username: string) /** :Promise<Boolean>*/ {
         try {
-            const ciao = await this.dao.deleteUser(username);
-            return ciao;
+            // Check if the logged-in user is not an admin and is trying to delete another user
+            if (user.role !== "Admin" && user.username !== username) {
+                throw new UserNotAdminError();
+            }
+            // Get the user to be deleted
+            const userToDelete = await this.dao.getUserByUsername(username);
+
+            // Check if the user to be deleted is an admin
+            if (userToDelete.role === "Admin") {
+                throw new UserIsAdminError();
+            }
+
+            // Proceed with deletion if checks pass
+            return this.dao.deleteUser(username);
         } catch (error) {
             throw error;
         }
-     }
+    }
 
     /**
      * Deletes all non-Admin users
@@ -90,8 +110,28 @@ class UserController {
      * @param username The username of the user to update. It must be equal to the username of the user parameter.
      * @returns A Promise that resolves to the updated user
      */
-    async updateUserInfo(user: User, name: string, surname: string, address: string, birthdate: string, username: string) /*:Promise<User>*/ { 
-        return this.dao.updateUserInfo(name, surname, address, birthdate, username)
+    async updateUserInfo(user: User, name: string, surname: string, address: string, birthdate: string, username: string) /*:Promise<User>*/ {
+        try {
+            if (user.role !== "Admin" && user.username !== username) {
+                throw new UserNotAdminError();
+            }
+            const userToEdit = await this.dao.getUserByUsername(username);
+
+            if (userToEdit.role === "Admin" && user.username !== username) {
+                throw new UnauthorizedEditError();
+            }
+
+            const bdate = new Date(birthdate);
+            const currentDate = new Date();
+            if (bdate > currentDate) {
+                throw new InvalidBirthdateError();
+            }
+
+            return this.dao.updateUserInfo(name, surname, address, birthdate, username)
+
+        } catch (error) {
+            throw error;
+        }
     }
 }
